Refactor project and workspace Effect services
