+++ apps/server/src/project/ProjectFaviconResolver.ts
@@ -1,13 +1,18 @@
+/**
+ * ProjectFaviconResolver - Effect service contract for project icon discovery.
+ *
+ * Resolves a representative favicon or app icon file for a workspace by
+ * checking common file locations and project source metadata.
+ *
+ * @module ProjectFaviconResolver
+ */
+import * as Context from "effect/Context";
 import * as Effect from "effect/Effect";
 import * as FileSystem from "effect/FileSystem";
 import * as Layer from "effect/Layer";
 import * as Path from "effect/Path";
 
-import {
-  ProjectFaviconResolver,
-  type ProjectFaviconResolverShape,
-} from "../Services/ProjectFaviconResolver.ts";
-import { WorkspacePaths } from "../../workspace/Services/WorkspacePaths.ts";
+import * as WorkspacePaths from "../workspace/WorkspacePaths.ts";
 
 // Well-known favicon paths checked in order.
 const FAVICON_CANDIDATES = [
@@ -51,6 +56,19 @@ const LINK_ICON_HTML_RE =
 const LINK_ICON_OBJ_RE =
   /(?=[^}]*\brel\s*:\s*["'](?:icon|shortcut icon)["'])(?=[^}]*\bhref\s*:\s*["']([^"'?]+))[^}]*/i;
 
+/** Service tag for project favicon resolution. */
+export class ProjectFaviconResolver extends Context.Service<
+  ProjectFaviconResolver,
+  {
+    /**
+     * Resolve a favicon or icon file path for the provided workspace root.
+     *
+     * Returns `null` when no candidate icon file can be found.
+     */
+    readonly resolvePath: (cwd: string) => Effect.Effect<string | null>;
+  }
+>()("t3/project/ProjectFaviconResolver") {}
+
 function extractIconHref(source: string): string | null {
   const htmlMatch = source.match(LINK_ICON_HTML_RE);
   if (htmlMatch?.[1]) return htmlMatch[1];
@@ -59,12 +77,12 @@ function extractIconHref(source: string): string | null {
   return null;
 }
 
-export const makeProjectFaviconResolver = Effect.gen(function* () {
+export const make = Effect.gen(function* () {
   const fileSystem = yield* FileSystem.FileSystem;
   const path = yield* Path.Path;
-  const workspacePaths = yield* WorkspacePaths;
+  const workspacePaths = yield* WorkspacePaths.WorkspacePaths;
 
-  const resolveIconHref = (href: string): string[] => {
+  const resolveIconHref = (href: string): ReadonlyArray<string> => {
     const clean = href.replace(/^\//, "");
     return [path.join("public", clean), clean];
   };
@@ -93,9 +111,9 @@ export const makeProjectFaviconResolver = Effect.gen(function* () {
     return null;
   });
 
-  const resolvePath: ProjectFaviconResolverShape["resolvePath"] = Effect.fn(
+  const resolvePath: ProjectFaviconResolver["Service"]["resolvePath"] = Effect.fn(
     "ProjectFaviconResolver.resolvePath",
-  )(function* (cwd: string): Effect.fn.Return<string | null> {
+  )(function* (cwd) {
     const projectCwd = yield* workspacePaths
       .normalizeWorkspaceRoot(cwd)
       .pipe(Effect.orElseSucceed(() => null));
@@ -138,12 +156,7 @@ export const makeProjectFaviconResolver = Effect.gen(function* () {
     return null;
   });
 
-  return {
-    resolvePath,
-  } satisfies ProjectFaviconResolverShape;
+  return ProjectFaviconResolver.of({ resolvePath });
 });
 
-export const ProjectFaviconResolverLive = Layer.effect(
-  ProjectFaviconResolver,
-  makeProjectFaviconResolver,
-);
+export const layer = Layer.effect(ProjectFaviconResolver, make);
