+++ apps/server/src/project/ProjectSetupScriptRunner.ts
@@ -1,0 +1,172 @@
+import * as Contracts from "@t3tools/contracts";
+import * as ProjectScripts from "@t3tools/shared/projectScripts";
+import * as Context from "effect/Context";
+import * as Effect from "effect/Effect";
+import * as Layer from "effect/Layer";
+import * as Option from "effect/Option";
+import * as Schema from "effect/Schema";
+
+import * as ProjectionSnapshotQuery from "../orchestration/Services/ProjectionSnapshotQuery.ts";
+import * as TerminalManager from "../terminal/Manager.ts";
+
+export interface ProjectSetupScriptRunnerResultNoScript {
+  readonly status: "no-script";
+}
+
+export interface ProjectSetupScriptRunnerResultStarted {
+  readonly status: "started";
+  readonly scriptId: string;
+  readonly scriptName: string;
+  readonly terminalId: string;
+  readonly cwd: string;
+}
+
+export type ProjectSetupScriptRunnerResult =
+  | ProjectSetupScriptRunnerResultNoScript
+  | ProjectSetupScriptRunnerResultStarted;
+
+export interface ProjectSetupScriptRunnerInput {
+  readonly threadId: string;
+  readonly projectId?: string;
+  readonly projectCwd?: string;
+  readonly worktreePath: string;
+  readonly preferredTerminalId?: string;
+}
+
+export class ProjectSetupScriptRunnerError extends Schema.TaggedErrorClass<ProjectSetupScriptRunnerError>()(
+  "ProjectSetupScriptRunnerError",
+  {
+    threadId: Schema.String,
+    projectId: Schema.optional(Schema.String),
+    projectCwd: Schema.optional(Schema.String),
+    worktreePath: Schema.String,
+    operation: Schema.String,
+    detail: Schema.String,
+    cause: Schema.optional(Schema.Defect()),
+  },
+) {
+  override get message(): string {
+    return `Project setup script failed in ${this.operation} for thread '${this.threadId}': ${this.detail}`;
+  }
+}
+
+export class ProjectSetupScriptRunner extends Context.Service<
+  ProjectSetupScriptRunner,
+  {
+    readonly runForThread: (
+      input: ProjectSetupScriptRunnerInput,
+    ) => Effect.Effect<ProjectSetupScriptRunnerResult, ProjectSetupScriptRunnerError>;
+  }
+>()("t3/project/ProjectSetupScriptRunner") {}
+
+const isProjectSetupScriptRunnerError = Schema.is(ProjectSetupScriptRunnerError);
+
+function detailFromUnknown(cause: unknown): string {
+  if (
+    typeof cause === "object" &&
+    cause !== null &&
+    "message" in cause &&
+    typeof cause.message === "string"
+  ) {
+    return cause.message;
+  }
+  return String(cause);
+}
+
+function runnerError(
+  input: ProjectSetupScriptRunnerInput,
+  operation: string,
+  detail: string,
+  cause?: unknown,
+): ProjectSetupScriptRunnerError {
+  return new ProjectSetupScriptRunnerError({
+    threadId: input.threadId,
+    worktreePath: input.worktreePath,
+    operation,
+    detail,
+    ...(input.projectId === undefined ? {} : { projectId: input.projectId }),
+    ...(input.projectCwd === undefined ? {} : { projectCwd: input.projectCwd }),
+    ...(cause === undefined ? {} : { cause }),
+  });
+}
+
+function mapRunnerError(input: ProjectSetupScriptRunnerInput, operation: string) {
+  return Effect.mapError((cause: unknown) =>
+    isProjectSetupScriptRunnerError(cause)
+      ? cause
+      : runnerError(input, operation, detailFromUnknown(cause), cause),
+  );
+}
+
+export const make = Effect.gen(function* () {
+  const projectionSnapshotQuery = yield* ProjectionSnapshotQuery.ProjectionSnapshotQuery;
+  const terminalManager = yield* TerminalManager.TerminalManager;
+
+  const runForThread: ProjectSetupScriptRunner["Service"]["runForThread"] = Effect.fn(
+    "ProjectSetupScriptRunner.runForThread",
+  )(function* (input) {
+    const projectById = input.projectId
+      ? yield* projectionSnapshotQuery
+          .getProjectShellById(Contracts.ProjectId.make(input.projectId))
+          .pipe(Effect.map(Option.getOrUndefined), mapRunnerError(input, "resolveProject"))
+      : null;
+    const project =
+      projectById ??
+      (input.projectCwd
+        ? yield* projectionSnapshotQuery
+            .getActiveProjectByWorkspaceRoot(input.projectCwd)
+            .pipe(Effect.map(Option.getOrUndefined), mapRunnerError(input, "resolveProject"))
+        : null);
+
+    if (!project) {
+      return yield* runnerError(
+        input,
+        "resolveProject",
+        "Project was not found for setup script execution.",
+      );
+    }
+
+    const script = ProjectScripts.setupProjectScript(project.scripts);
+    if (!script) {
+      return {
+        status: "no-script",
+      } as const;
+    }
+
+    const terminalId = input.preferredTerminalId ?? `setup-${script.id}`;
+    const cwd = input.worktreePath;
+    const env = ProjectScripts.projectScriptRuntimeEnv({
+      project: { cwd: project.workspaceRoot },
+      worktreePath: input.worktreePath,
+    });
+
+    yield* terminalManager
+      .open({
+        threadId: input.threadId,
+        terminalId,
+        cwd,
+        worktreePath: input.worktreePath,
+        env,
+      })
+      .pipe(mapRunnerError(input, "openTerminal"));
+    yield* terminalManager
+      .write({
+        threadId: input.threadId,
+        terminalId,
+        data: `${script.command}\r`,
+      })
+      .pipe(mapRunnerError(input, "writeCommand"));
+
+    return {
+      status: "started",
+      scriptId: script.id,
+      scriptName: script.name,
+      terminalId,
+      cwd,
+    } as const;
+  });
+
+  return ProjectSetupScriptRunner.of({ runForThread });
+});
+
+export const layer = Layer.effect(ProjectSetupScriptRunner, make);
