+++ apps/server/src/project/Services/RepositoryIdentityResolver.ts
@@ -1,12 +1,2 @@
-import type { RepositoryIdentity } from "@t3tools/contracts";
-import * as Context from "effect/Context";
-import type * as Effect from "effect/Effect";
-
-export interface RepositoryIdentityResolverShape {
-  readonly resolve: (cwd: string) => Effect.Effect<RepositoryIdentity | null>;
-}
-
-export class RepositoryIdentityResolver extends Context.Service<
-  RepositoryIdentityResolver,
-  RepositoryIdentityResolverShape
->()("t3/project/Services/RepositoryIdentityResolver") {}
+// Compatibility export for orchestration modules excluded by #2829.
+export * from "../RepositoryIdentityResolver.ts";
