+++ apps/server/src/workspace/WorkspaceSearchIndex.ts
@@ -1,4 +1,4 @@
-import { FileFinder, type MixedItem, type MixedSearchResult } from "@ff-labs/fff-node";
+import * as FFF from "@ff-labs/fff-node";
 import * as Context from "effect/Context";
 import * as Effect from "effect/Effect";
 import * as Layer from "effect/Layer";
@@ -6,11 +6,7 @@ import * as LayerMap from "effect/LayerMap";
 import * as Schedule from "effect/Schedule";
 import * as Schema from "effect/Schema";
 
-import type {
-  ProjectEntry,
-  ProjectListEntriesResult,
-  ProjectSearchEntriesResult,
-} from "@t3tools/contracts";
+import type * as Contracts from "@t3tools/contracts";
 
 const WORKSPACE_INDEX_MAX_ENTRIES = 25_000;
 const WORKSPACE_INDEX_PAGE_SIZE = WORKSPACE_INDEX_MAX_ENTRIES + 2;
@@ -75,11 +71,14 @@ export type WorkspaceSearchIndexError =
 export class WorkspaceSearchIndex extends Context.Service<
   WorkspaceSearchIndex,
   {
-    readonly list: () => Effect.Effect<ProjectListEntriesResult, WorkspaceSearchIndexSearchFailed>;
+    readonly list: () => Effect.Effect<
+      Contracts.ProjectListEntriesResult,
+      WorkspaceSearchIndexSearchFailed
+    >;
     readonly search: (
       query: string,
       limit: number,
-    ) => Effect.Effect<ProjectSearchEntriesResult, WorkspaceSearchIndexSearchFailed>;
+    ) => Effect.Effect<Contracts.ProjectSearchEntriesResult, WorkspaceSearchIndexSearchFailed>;
     readonly refresh: () => Effect.Effect<
       void,
       WorkspaceSearchIndexRefreshFailed | WorkspaceSearchIndexScanTimedOut
@@ -100,7 +99,7 @@ function parentPathOf(input: string): string | undefined {
   return separatorIndex === -1 ? undefined : input.slice(0, separatorIndex);
 }
 
-function toProjectEntry(item: MixedItem): ProjectEntry | null {
+function toProjectEntry(item: FFF.MixedItem): Contracts.ProjectEntry | null {
   const normalizedPath = trimDirectorySeparator(toPosixPath(item.item.relativePath));
   if (!normalizedPath) {
     return null;
@@ -113,10 +112,10 @@ function toProjectEntry(item: MixedItem): ProjectEntry | null {
 }
 
 function mapMixedSearchResult(
-  result: MixedSearchResult,
+  result: FFF.MixedSearchResult,
   limit: number,
-): { readonly entries: ProjectEntry[]; readonly truncated: boolean } {
-  const entries: ProjectEntry[] = [];
+): { readonly entries: Contracts.ProjectEntry[]; readonly truncated: boolean } {
+  const entries: Contracts.ProjectEntry[] = [];
   for (const item of result.items) {
     const entry = toProjectEntry(item);
     if (entry) {
@@ -138,7 +137,9 @@ function mapMixedSearchResult(
   };
 }
 
-function withDirectoryAncestors(entries: ReadonlyArray<ProjectEntry>): ProjectEntry[] {
+function withDirectoryAncestors(
+  entries: ReadonlyArray<Contracts.ProjectEntry>,
+): Contracts.ProjectEntry[] {
   const entryByPath = new Map(entries.map((entry) => [entry.path, entry]));
   for (const entry of entries) {
     let parentPath = parentPathOf(entry.path);
@@ -153,7 +154,7 @@ function withDirectoryAncestors(entries: ReadonlyArray<ProjectEntry>): ProjectEn
 }
 
 const createFinder = Effect.fn("WorkspaceSearchIndex.createFinder")(function* (cwd: string) {
-  const result = FileFinder.create({
+  const result = FFF.FileFinder.create({
     basePath: cwd,
     disableMmapCache: true,
     disableContentIndexing: true,
@@ -167,7 +168,7 @@ const createFinder = Effect.fn("WorkspaceSearchIndex.createFinder")(function* (c
 
 const waitForScan = Effect.fn("WorkspaceSearchIndex.waitForScan")(function* (
   cwd: string,
-  finder: FileFinder,
+  finder: FFF.FileFinder,
 ) {
   yield* Effect.sync(() => finder.isScanning()).pipe(
     Effect.repeat({
@@ -182,64 +183,69 @@ const waitForScan = Effect.fn("WorkspaceSearchIndex.waitForScan")(function* (
   );
 });
 
-const makeWorkspaceSearchIndex = (cwd: string) =>
-  Effect.acquireRelease(createFinder(cwd), (finder) => Effect.sync(() => finder.destroy())).pipe(
-    Effect.tap((finder) => waitForScan(cwd, finder)),
-    Effect.map((finder) => {
-      const runMixedSearch = Effect.fn("WorkspaceSearchIndex.runMixedSearch")(function* (
-        query: string,
-        pageSize: number,
-      ) {
-        const result = yield* Effect.sync(() => finder.mixedSearch(query, { pageSize }));
-        if (!result.ok) {
-          return yield* new WorkspaceSearchIndexSearchFailed({ cwd, reason: result.error });
-        }
-        return result.value;
-      });
+export const make = Effect.fn("WorkspaceSearchIndex.make")(function* (cwd: string) {
+  const finder = yield* Effect.acquireRelease(createFinder(cwd), (finder) =>
+    Effect.sync(() => finder.destroy()),
+  );
+  yield* waitForScan(cwd, finder);
 
-      const refresh: WorkspaceSearchIndex["Service"]["refresh"] = Effect.fn(
-        "WorkspaceSearchIndex.refresh",
-      )(function* () {
-        const result = yield* Effect.sync(() => finder.scanFiles());
-        if (!result.ok) {
-          return yield* new WorkspaceSearchIndexRefreshFailed({ cwd, reason: result.error });
-        }
-        yield* waitForScan(cwd, finder);
-      });
+  const runMixedSearch = Effect.fn("WorkspaceSearchIndex.runMixedSearch")(function* (
+    query: string,
+    pageSize: number,
+  ) {
+    const result = yield* Effect.sync(() => finder.mixedSearch(query, { pageSize }));
+    if (!result.ok) {
+      return yield* new WorkspaceSearchIndexSearchFailed({ cwd, reason: result.error });
+    }
+    return result.value;
+  });
 
-      const list: WorkspaceSearchIndex["Service"]["list"] = Effect.fn("WorkspaceSearchIndex.list")(
-        function* () {
-          const result = yield* runMixedSearch("", WORKSPACE_INDEX_PAGE_SIZE);
-          const mapped = mapMixedSearchResult(result, WORKSPACE_INDEX_MAX_ENTRIES);
-          const sortedEntries = withDirectoryAncestors(mapped.entries).toSorted((left, right) =>
-            left.path.localeCompare(right.path),
-          );
-          const entries = sortedEntries.slice(0, WORKSPACE_INDEX_MAX_ENTRIES);
-          return {
-            entries,
-            truncated: mapped.truncated || entries.length < sortedEntries.length,
-          };
-        },
+  const refresh: WorkspaceSearchIndex["Service"]["refresh"] = Effect.fn(
+    "WorkspaceSearchIndex.refresh",
+  )(function* () {
+    const result = yield* Effect.sync(() => finder.scanFiles());
+    if (!result.ok) {
+      return yield* new WorkspaceSearchIndexRefreshFailed({ cwd, reason: result.error });
+    }
+    yield* waitForScan(cwd, finder);
+  });
+
+  const list: WorkspaceSearchIndex["Service"]["list"] = Effect.fn("WorkspaceSearchIndex.list")(
+    function* () {
+      const result = yield* runMixedSearch("", WORKSPACE_INDEX_PAGE_SIZE);
+      const mapped = mapMixedSearchResult(result, WORKSPACE_INDEX_MAX_ENTRIES);
+      const sortedEntries = withDirectoryAncestors(mapped.entries).toSorted((left, right) =>
+        left.path.localeCompare(right.path),
       );
-
-      const search: WorkspaceSearchIndex["Service"]["search"] = Effect.fn(
-        "WorkspaceSearchIndex.search",
-      )(function* (query, limit) {
-        const result = yield* runMixedSearch(query, Math.max(1, limit + 1));
-        return mapMixedSearchResult(result, limit);
-      });
-
-      return WorkspaceSearchIndex.of({ list, refresh, search });
-    }),
+      const entries = sortedEntries.slice(0, WORKSPACE_INDEX_MAX_ENTRIES);
+      return {
+        entries,
+        truncated: mapped.truncated || entries.length < sortedEntries.length,
+      };
+    },
   );
 
-const workspaceSearchIndexLayer = (cwd: string) =>
-  Layer.effect(WorkspaceSearchIndex, makeWorkspaceSearchIndex(cwd));
+  const search: WorkspaceSearchIndex["Service"]["search"] = Effect.fn(
+    "WorkspaceSearchIndex.search",
+  )(function* (query, limit) {
+    const result = yield* runMixedSearch(query, Math.max(1, limit + 1));
+    return mapMixedSearchResult(result, limit);
+  });
+
+  return WorkspaceSearchIndex.of({ list, refresh, search });
+});
+
+/**
+ * A layer factory is required because every index is scoped to a concrete
+ * workspace root. WorkspaceSearchIndexMap owns memoization and idle cleanup;
+ * using a default cwd here would mix resources from different workspaces.
+ */
+export const layer = (cwd: string) => Layer.effect(WorkspaceSearchIndex, make(cwd));
 
 export class WorkspaceSearchIndexMap extends LayerMap.Service<WorkspaceSearchIndexMap>()(
   "t3/workspace/WorkspaceSearchIndexMap",
   {
-    lookup: workspaceSearchIndexLayer,
+    lookup: layer,
     idleTimeToLive: WORKSPACE_INDEX_IDLE_TTL,
   },
 ) {}
